Add optional splitBy prop to RouteForm and refactor component
A new optional property, `splitBy`, was added to the `RouteForm` component. This allows consumers of the `RouteForm` component to pass in a default value for the `splitBy` state. The existing `splitBy` state variable was renamed to `splitByValue` to avoid naming conflicts.

Additionally, useEffect has been added to update `splitByValue` whenever `splitBy` prop changes. This ensures that if the `splitBy` prop is updated externally, the `RouteForm` reflects this change. Corrections are also made in hook dependencies and the onAnalyzeRouteClick method, using `splitByValue` instead of `splitBy`. This makes the component more flexible and less prone to errors due to naming confusion.